Development

+++ code/modules/mob/living/carbon/brain/brain_item.dm
@@ -33,6 +33,9 @@
 	icon = 'icons/mob/alien.dmi'
 	icon_state = "chitin"
 
+/obj/item/organ/brain/xeno/gain_trauma()
+	return
+	
 /obj/item/organ/brain/Initialize(mapload)
 	. = ..()
 	health = config.default_brain_health
@@ -188,8 +191,11 @@
 	if(trauma && (cure_permanent || !trauma.permanent))
 		qdel(trauma)
 
-/obj/item/organ/brain/proc/cure_all_traumas(cure_permanent = FALSE)
+/obj/item/organ/brain/proc/cure_all_traumas(cure_permanent = FALSE, cure_type = "")
 	for(var/X in traumas)
 		var/datum/brain_trauma/trauma = X
-		if(cure_permanent || !trauma.permanent)
-			qdel(trauma)
+		if(trauma.cure_type == cure_type || cure_type == CURE_ADMIN)
+			if(cure_permanent || !trauma.permanent)
+				qdel(trauma)
+				if(cure_type != CURE_ADMIN)
+					break
